Create README.md
At this point I’ve learned the basics fundamentals, but haven’t put much code together. Give a little data and hopefully collaborate with experienced coders and work my way up ladder of knowledge and skills to get a better understanding of the value of efficient code. Thanks to the mystery people that got me here and hopefully prosperous collaboration in the business of data and building a business relationship that thrives.